move LogoFull outside icons folder
