+++ components/layout/MobileNavBar.tsx
@@ -4,7 +4,7 @@ import styles from './MobileNavBar.module.css';
 import menuStyles from './MobileNavMenu.module.css';
 import Exit from '../icons-v2/Exit';
 import Hamburger from '../icons-v2/Hamburger';
-import LogoFull from '../icons/LogoFull';
+import LogoFull from '../LogoFull';
 import PortfolioLogo from '../logos/PortfolioLogo';
 
 interface TabProps {
